src/main/java/com/greenfoxacademy/springwebapp/kingdom/controllers/KingdomController.java
package com.greenfoxacademy.springwebapp.kingdom.controllers;

import com.greenfoxacademy.springwebapp.globalexceptionhandling.IdNotFoundException;
import com.greenfoxacademy.springwebapp.kingdom.models.KingdomEntity;
import com.greenfoxacademy.springwebapp.kingdom.services.KingdomService;
import lombok.AllArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@AllArgsConstructor
@RequestMapping(KingdomController.KINGDOM)
public class KingdomController {

  public static final String KINGDOM = "/kingdom";

  private final KingdomService kingdomService;

  @GetMapping("/{id}")
  public ResponseEntity<Object> getKingdomByID(@PathVariable Long id) throws IdNotFoundException {
    KingdomEntity kingdom = kingdomService.findByID(id);
    return ResponseEntity.ok(kingdomService.entityToKingdomResponseDTO(kingdom));
  }
}